Fixing landing page
