+++ website/templates/blog/index.html
@@ -9,3 +9,7 @@
   </section>
 </div><!-- /#content -->
 {% endblock content %}
+
+
+
+
