+++ website/templates/footer.html
@@ -1,7 +1,4 @@
 
-
-<footer class="footer">
-    <div class="text-center">
-      <p>Build with Flask and <span class="love">❤</span>. Get the source<a href="https://github.com/vcaldas/vcaldas.github.io"> here!</a> 
-    </div>
-</footer>
+<div class="text-center">
+  <p>Build with Flask and <span class="love">❤</span>. Get the source<a href="https://github.com/vcaldas/vcaldas.github.io"> here!</a> 
+</div>
