+++ website/templates/landing.html
@@ -1,43 +1,28 @@
 <!DOCTYPE html>
 <html lang="en">
+  <head>
+    {%include "head.html" %}
+  </head>
 
-<head>
-  {%include "head.html" %}
-
-</head>
-
-<body id="page-top">
-  <div id="particles-js">
-    <!-- Navigation -->
-    {% include "navbar/navigation.html" %}
-    <div class="content">
-      <h1>
-        <span class="site-title">V/Caldas</span>
-        <span class="site-description"></span>
-      </h1>
-      <div class="icons-header">
-        <a aria-label="Send email" href="mailto:{{config.email}}"><i class="icon fa fa-envelope"></i></a>
-        <a aria-label="My Research" href="/publications"><i class="icon fa fa-flask" aria-hidden="true"></i></a>
-        <a aria-label="My Vimeo" target="_blank" href="{{config.vimeo}}"><i class="icon fa fa-vimeo"
-            aria-hidden="true"></i></a>
-        <a aria-label="My Github" target="_blank" href="https://github.com/{{config.github}}"><i
-            class="icon fa fa-github-alt" aria-hidden="true"></i></a>
-      </div>
-    </div>
-    <a class="down" href="#contact" data-scroll><i class="icon fa fa-chevron-down" aria-hidden="true"></i></a>
-
-
+  <body>
+    <div id="particles-js">
+      <!-- Navigation -->
+      {% include "navbar/navigation.html" %}
+      {% include "landing/icon_row.html" %}
+      <a class="down" href="#contact" data-scroll>
+        <i class="icon fa fa-chevron-down" aria-hidden="true"></i>
+      </a>
   </div>
+    {% include "landing/contactrow.html" %}
 
-  <!-- About Section -->
-  {% include "contactrow.html" %}
+  <div id="footer">
+    <footer>
+      {% include "footer.html" %}
+    </footer>
+  </div>
+    
+    {% include "scripts.html" %}
 
-</body>
-
-<!-- Footer -->
-{% include "footer.html" %}
-
-<!-- javascript -->
-{% include "scripts.html" %}
-
+  </body>
 </html>
+
