+++ website/templates/landing/icon_row.html
@@ -1,0 +1,14 @@
+<div class="content">
+    <h1>
+      <span class="site-title">V/Caldas</span>
+      <span class="site-description"></span>
+    </h1>
+    <div class="icons-header">
+      <a aria-label="Send email" href="mailto:{{config.email}}"><i class="icon fa fa-envelope"></i></a>
+      <a aria-label="My Research" href="/publications"><i class="icon fa fa-flask" aria-hidden="true"></i></a>
+      <a aria-label="My Vimeo" target="_blank" href="{{config.vimeo}}"><i class="icon fa fa-vimeo"
+          aria-hidden="true"></i></a>
+      <a aria-label="My Github" target="_blank" href="https://github.com/{{config.github}}"><i
+          class="icon fa fa-github-alt" aria-hidden="true"></i></a>
+    </div>
+</div>
